Introduce SentryEvent::add_exception()
Replaces InternalSDK::capture_error().
