+++ src/sentry/native/native_event.h
@@ -47,6 +47,8 @@ public:
 	virtual void remove_tag(const String &p_key) override;
 	virtual String get_tag(const String &p_key) override;
 
+	virtual void add_exception(const String &p_type, const String &p_value, const Vector<StackFrame> &p_frames) override;
+
 	NativeEvent(sentry_value_t p_event);
 	NativeEvent();
 	virtual ~NativeEvent() override;
