+++ src/sentry/native/native_sdk.cpp
@@ -317,44 +317,6 @@ String NativeSDK::get_last_event_id() {
 	return _uuid_as_string(last_uuid);
 }
 
-String NativeSDK::capture_error(const String &p_type, const String &p_value, Level p_level, const Vector<StackFrame> &p_frames) {
-	sentry_value_t event = sentry_value_new_event();
-	sentry_value_set_by_key(event, "level",
-			sentry_value_new_string(sentry::level_as_cstring(p_level)));
-
-	sentry_value_t exception = sentry_value_new_exception(p_type.utf8(), p_value.utf8());
-	sentry_value_t stack_trace = sentry_value_new_object();
-
-	sentry_value_t frames = sentry_value_new_list();
-	for (const StackFrame &frame : p_frames) {
-		sentry_value_t sentry_frame = sentry_value_new_object();
-		sentry_value_set_by_key(sentry_frame, "filename", sentry_value_new_string(frame.filename.utf8()));
-		sentry_value_set_by_key(sentry_frame, "function", sentry_value_new_string(frame.function.utf8()));
-		sentry_value_set_by_key(sentry_frame, "lineno", sentry_value_new_int32(frame.lineno));
-		sentry_value_set_by_key(sentry_frame, "in_app", sentry_value_new_bool(frame.in_app));
-		sentry_value_set_by_key(sentry_frame, "platform", sentry_value_new_string(frame.platform.utf8()));
-		if (!frame.context_line.is_empty()) {
-			sentry_value_set_by_key(sentry_frame, "context_line", sentry_value_new_string(frame.context_line.utf8()));
-			sentry_value_set_by_key(sentry_frame, "pre_context", sentry::native::strings_to_sentry_list(frame.pre_context));
-			sentry_value_set_by_key(sentry_frame, "post_context", sentry::native::strings_to_sentry_list(frame.post_context));
-		}
-		if (frame.vars.size() > 0) {
-			sentry_value_t vars = sentry_value_new_object();
-			sentry_value_set_by_key(sentry_frame, "vars", vars);
-			for (auto pair : frame.vars) {
-				sentry_value_set_by_key(vars, pair.first.utf8(), sentry::native::variant_to_sentry_value(pair.second));
-			}
-		}
-		sentry_value_append(frames, sentry_frame);
-	}
-
-	sentry_value_set_by_key(stack_trace, "frames", frames);
-	sentry_value_set_by_key(exception, "stacktrace", stack_trace);
-	sentry_event_add_exception(event, exception);
-	last_uuid = sentry_capture_event(event);
-	return _uuid_as_string(last_uuid);
-}
-
 Ref<SentryEvent> NativeSDK::create_event() {
 	sentry_value_t event_value = sentry_value_new_event();
 	Ref<SentryEvent> event = memnew(NativeEvent(event_value));
